[rag] - extract SearchResult to retrieval/results.py
